Changed focus inset per Orion's review, splitting up ToggleButton and ToggleButtonGroup, added appropriate Storybook, tests, and docs to each.
