+++ packages/react-magma-dom/src/components/ToggleButton/ToggleButton.stories.tsx
@@ -1,24 +1,12 @@
 import React from 'react';
-import {
-  CheckIcon,
-  FormatAlignCenterIcon,
-  FormatAlignJustifyIcon,
-  FormatAlignLeftIcon,
-  FormatAlignRightIcon,
-  SettingsIcon,
-} from 'react-magma-icons';
+import { CheckIcon } from 'react-magma-icons';
 import { Meta, Story } from '@storybook/react/types-6-0';
-import { ToggleButtonGroup, ToggleButton, ToggleButtonGroupProps } from '.';
+import { ToggleButton, ToggleButtonProps } from '.';
 import { Container } from '../Container';
 import { ButtonSize } from '../Button';
 
-const Template: Story<ToggleButtonGroupProps> = args => (
-  <ToggleButtonGroup {...args}>
-    <ToggleButton aria-label="Check icon" icon={<CheckIcon />} />
-    <ToggleButton disabled aria-label="Check icon" icon={<CheckIcon />} />
-    <ToggleButton aria-label="Check icon" defaultChecked icon={<CheckIcon />} />
-    <ToggleButton aria-label="Check icon" icon={<CheckIcon />} />
-  </ToggleButtonGroup>
+const Template: Story<ToggleButtonProps> = args => (
+  <ToggleButton {...args} aria-label="Check icon" icon={<CheckIcon />} />
 );
 
 export default {
@@ -32,22 +20,12 @@ export default {
     ),
   ],
   argTypes: {
+    disabled: {
+      control: {
+        type: 'boolean',
+      },
+    },
     isInverse: {
-      control: {
-        type: 'boolean',
-      },
-    },
-    noSpace: {
-      control: {
-        type: 'boolean',
-      },
-    },
-    requiredSelect: {
-      control: {
-        type: 'boolean',
-      },
-    },
-    singleSelect: {
       control: {
         type: 'boolean',
       },
@@ -64,66 +42,4 @@ export default {
 export const Default = Template.bind({});
 Default.args = {
   isInverse: false,
-};
-
-export const SingleButton = args => {
-  return (
-    <ToggleButton
-      {...args}
-      aria-label="Settings icon"
-      icon={<SettingsIcon />}
-      size={ButtonSize.large}
-    />
-  );
-};
-SingleButton.args = {
-  ...Default.args,
-};
-
-export const SingleTextButton = args => {
-  return (
-    <ToggleButton {...args} size={ButtonSize.small}>
-      Hello
-    </ToggleButton>
-  );
-};
-SingleTextButton.args = {
-  ...Default.args,
-};
-
-export const AlignmentExample = args => {
-  return (
-    <ToggleButtonGroup {...args}>
-      <ToggleButton aria-label="Left align" icon={<FormatAlignLeftIcon />} />
-      <ToggleButton
-        aria-label="Center align"
-        icon={<FormatAlignCenterIcon />}
-      />
-      <ToggleButton aria-label="Right align" icon={<FormatAlignRightIcon />} />
-      <ToggleButton
-        aria-label="Justify align"
-        icon={<FormatAlignJustifyIcon />}
-      />
-    </ToggleButtonGroup>
-  );
-};
-AlignmentExample.args = {
-  noSpace: true,
-  requiredSelect: true,
-  singleSelect: true,
-  ...Default.args,
-};
-
-export const DifferentToggleButtons = args => {
-  return (
-    <ToggleButtonGroup {...args}>
-      <ToggleButton aria-label="Settings icon" icon={<SettingsIcon />} />
-      <ToggleButton>Text</ToggleButton>
-      <ToggleButton icon={<SettingsIcon />}>Icon and Text</ToggleButton>
-    </ToggleButtonGroup>
-  );
-};
-DifferentToggleButtons.args = {
-  noSpace: true,
-  ...Default.args,
 };
