+++ packages/react-magma-dom/src/components/ToggleButton/index.ts
@@ -1,2 +1,1 @@
 export * from './ToggleButton';
-export * from './ToggleButtonGroup';
