+++ packages/react-magma-dom/src/index.ts
@@ -245,3 +245,4 @@ export * from './components/ButtonGroup';
 
 export * from './components/CharacterCounter';
 export * from './components/ToggleButton';
+export * from './components/ToggleButtonGroup';
